backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
social-auth-app-django==5.4.0
social-auth-core==4.5.4
Pillow==10.1.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
razorpay==1.3.0
django-ratelimit==4.1.0
boto3==1.28.0
django-storages==1.14.2
minio==7.2.0
